+++ src/components/Form/Form.tsx
@@ -1,6 +1,8 @@
 import * as React from 'react';
 import {autobind} from '@shopify/javascript-utilities/decorators';
 
+import * as styles from './Form.scss';
+
 export type Enctype =
   | 'application/x-www-form-urlencoded'
   | 'multipart/form-data'
@@ -9,7 +11,7 @@ export type Method = 'post' | 'get' | 'action';
 export type Target = '_blank' | '_self' | '_parent' | '_top' | string;
 
 export interface Props {
-  /** Space seperated list of character encodings */
+  /** Space separated list of character encodings */
   acceptCharset?: string;
   /** Where to send form-data on submittal */
   action?: string;
@@ -19,11 +21,13 @@ export interface Props {
   children?: React.ReactNode;
   /** Media type when submiting content to server */
   encType?: Enctype;
+  /** Toggles if form submits on Enter keypress. Defaults to true. */
+  implicitSubmit?: boolean;
   /** Method used to submit form */
   method?: Method;
   /** A unique name for the form */
   name?: string;
-  /** Wheather or not form is validated when submitting */
+  /** Whether or not form is validated when submitting */
   noValidate?: boolean;
   /** Blocks the default form action */
   preventDefault?: boolean;
@@ -41,12 +45,17 @@ export default class Form extends React.PureComponent<Props> {
       autoComplete,
       children,
       encType,
+      implicitSubmit = true,
       method,
       name,
       noValidate,
       target,
     } = this.props;
     const autoCompleteInputs = normalizeAutoComplete(autoComplete);
+
+    const submitMarkup = implicitSubmit && (
+      <button type="submit" className={styles.submit} aria-hidden="true" />
+    );
 
     return (
       <form
@@ -61,6 +70,7 @@ export default class Form extends React.PureComponent<Props> {
         onSubmit={this.handleSubmit}
       >
         {children}
+        {submitMarkup}
       </form>
     );
   }